moved the ImageReverser to qrcode component folder
