+++ frontend/src/components/auth/2fa/otp/qrcode/ImageReverser.tsx
